Implement SRTCP.
Add SRTCP context to process SRTCP packets. Add necessary stuff to ZRTP and the transformers to enable the SRTCP transport
The interoperability with libsrtp was tested in off-line mode using reference data, see SrtcpTest in the demo directory.
libsrtp does not provide an online test program to test SRTCP interoperability.